# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bhakti App API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for Bhakti App - A spiritual services booking platform",
    "contact": {
      "name": "API Support",
      "email": "support@bhaktiapp.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "https://api.bhaktiapp.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint. Can be provided via Authorization header or cookie."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT token stored in httpOnly cookie"
      }
    },
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total number of items"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "description": "Items per page"
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages"
          }
        }
      },
...
      "CoreEvent": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "ObjectId"
          },
          "type": {
            "type": "string",
            "enum": [
              "coreevent_online_puja",
              "coreevent_offline_puja"
            ],
            "example": "coreevent_online_puja"
          },
          "title": {
            "type": "string",
            "example": "Online Puja"
          },
          "description": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "shadowColor": {
            "type": "string"
          },
          "visible": {
            "type": "boolean",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "ObjectId"
          },
          "coreEventId": {
            "type": "string",
            "example": "coreevent_online_puja"
          },
          "eventName": {
            "type": "string",
            "example": "Ganesh Chaturthi Puja"
          },
          "templeId": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "ObjectId"
            }
          },
          "packageId": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "ObjectId"
            }
          },
          "pricePackageId": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "packageId": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "discount": {
                  "type": "number"
                }
              }
            }
          },
          "eventStartTime": {
            "type": "string",
            "format": "date-time"
          },
          "eventExpirationTime": {
            "type": "string",
            "format": "date-time"
          },
          "isPopular": {
            "type": "boolean",
            "default": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/api/coreevent/create": {
      "post": {
        "summary": "Create a new core event",
        "tags": [
          "Core Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "title",
                  "icon"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "coreevent_online_puja",
                      "coreevent_offline_puja"
                    ],
                    "example": "coreevent_online_puja"
                  },
                  "title": {
                    "type": "string",
                    "example": "Online Puja"
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  },
                  "shadowColor": {
                    "type": "string"
                  },
                  "visible": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Core event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreEvent"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or duplicate core event type"
          }
        }
      }
    },
    "/api/coreevent/update": {
      "patch": {
        "summary": "Update a core event",
        "tags": [
          "Core Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "coreevent_online_puja",
                      "coreevent_offline_puja"
                    ],
                    "description": "Core event type (used to identify the event)"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  },
                  "shadowColor": {
                    "type": "string"
                  },
                  "visible": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Core event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreEvent"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid core event type"
          }
        }
      }
    },
    "/api/coreevent/get": {
      "get": {
        "summary": "Get all core events with pagination",
        "tags": [
          "Core Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of core events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CoreEvent"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/coreevent/getByTypes": {
      "post": {
        "summary": "Get core events by types",
        "tags": [
          "Core Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "types"
                ],
                "properties": {
                  "types": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "coreevent_online_puja",
                        "coreevent_offline_puja"
                      ]
                    },
                    "description": "Array of core event types"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of core events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoreEvent"
                  }
                }
              }
            }
          },
          "400": {
            "description": "types must be a non-empty array"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/event/create": {
      "post": {
        "summary": "Create a new event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "coreEventId",
                  "eventName",
                  "templeId",
                  "packageId",
                  "pricePackageId",
                  "eventStartTime",
                  "eventExpirationTime"
                ],
                "properties": {
                  "coreEventId": {
                    "type": "string",
                    "example": "coreevent_online_puja"
                  },
                  "eventName": {
                    "type": "string",
                    "example": "Ganesh Chaturthi Puja"
                  },
                  "templeId": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "ObjectId"
                    }
                  },
                  "packageId": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "ObjectId"
                    }
                  },
                  "pricePackageId": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "packageId": {
                          "type": "string"
                        },
                        "price": {
                          "type": "number"
                        },
                        "discount": {
                          "type": "number"
                        }
                      }
                    }
                  },
                  "eventStartTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "eventExpirationTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "isPopular": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/event/update": {
      "patch": {
        "summary": "Update an event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "ObjectId",
                    "description": "Event ID"
                  },
                  "coreEventId": {
                    "type": "string"
                  },
                  "eventName": {
                    "type": "string"
                  },
                  "templeId": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "packageId": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "pricePackageId": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "eventStartTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "eventExpirationTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "isPopular": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Invalid event id or validation error"
          }
        }
      }
    },
    "/api/event/delete": {
      "delete": {
        "summary": "Delete an event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "ObjectId",
                    "description": "Event ID to delete"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid event id"
          }
        }
      }
    },
    "/api/event/get": {
      "get": {
        "summary": "Get all events with pagination and filters",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10000
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "templeId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by temple ID"
          },
          {
            "in": "query",
            "name": "coreEventId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by core event ID"
          },
          {
            "in": "query",
            "name": "packageId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by package ID"
          },
          {
            "in": "query",
            "name": "isPopular",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by popular status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Event"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}